% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnetsurvfit.R
\name{breslow}
\alias{breslow}
\title{Compute Breslow estimates of survival functions}
\usage{
breslow(fit, centered = FALSE)
}
\arguments{
\item{fit}{fitted model objects with the objects described in details section.}

\item{centered}{if \code{TRUE} (default), return data from a predicted survival function at the mean values of the predictors, if \code{FALSE} returns prediction for all predictors equal to zero (baseline hazard).}
}
\description{
Compute Breslow estimates of survival functions
}
\details{
This function computes the Breslow's survival functions. It requires a model object with the \code{y}, a \code{Surv} object, estimated model coefficients, \code{coefficients} and the covariate mean values (estimate from model.matrix), \code{means}.
}
