% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnetsurvfit.R
\name{concordScore}
\alias{concordScore}
\title{Compute the concordance statistic for the pcoxtime model}
\usage{
concordScore(fit, newdata = NULL, stats = FALSE)
}
\arguments{
\item{fit}{fitted \code{\link[glmnetsurv]{glmnetsurv}}, \code{\link[pcoxtime]{pcoxtime}}, \code{\link[survival]{coxph}}, etc.}

\item{newdata}{optional data frame containing the variables appearing on the right hand side of \code{\link[glmnetsurv]{glmnetsurv}} formula.}

\item{stats}{logical. If \code{TRUE} all the related concordance statistics are returned.}
}
\value{
an object containing the concordance, followed by the number of pairs that agree, disagree, are tied, and are not comparable.
}
\description{
The function computes the agreement between the observed response and the predictor.
}
\details{
Computes Harrel's C index for predictions for \code{\link[glmnetsurv]{glmnetsurv}}, \code{\link[pcoxtime]{pcoxtime}}, \code{\link[survival]{coxph}}, etc, object and takes into account censoring. See \code{\link[survival]{survConcordance}}.
}
\examples{

data(veteran, package="survival")
# Penalized
lam <- 0.1
alp <- 0.5
pfit1 <- glmnetsurv(Surv(time, status) ~ factor(trt) + karno + diagtime + age + prior
	, data = veteran
	, lambda = lam
	, alpha = alp
)
c1 <- concordScore(pfit1)
c1

# Unpenalized
lam <- 0
alp <- 1
pfit2 <- glmnetsurv(Surv(time, status) ~ factor(trt) + karno + diagtime + age + prior
	, data = veteran
	, lambda = lam
	, alpha = alp
)
c2 <- concordScore(pfit2)
c2

}
