% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnetpostsurv.R
\docType{package}
\name{glmnetpostsurv}
\alias{glmnetpostsurv}
\title{Perform post fitting procedures to glmnet survival model}
\description{
Extends functionality of survival models in \code{\link[glmnet]{glmnet}} to compute survival curves and other calibrations.
}
