% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnetpostsurv.R
\docType{package}
\name{glmnetsurv}
\alias{glmnetsurv}
\title{Fit and perform post fitting procedures to glmnet survival models}
\usage{
glmnetsurv(
  formula = formula(data),
  data = sys.parent(),
  fittype = c("fit", "cv"),
  family = "cox",
  alpha = 1,
  lambda = NULL,
  contrasts.arg = NULL,
  xlevs = NULL,
  na.action = na.omit,
  ...
)
}
\arguments{
\item{formula}{Object of class formula describing 
the model. The response and terms are specified 
similar to \code{\link[survival]{Surv}} function.}

\item{data}{optional data frame containing 
variables specified in the formula.}

\item{fittype}{whether to perform cross-validation \code{(fittype = "cv")} or fit a \code{glmnet} model \code{(fittype = "fit")}.}

\item{family}{currently, only \code{glmnet} \code{"cox"} family (survival model) is allowed.}

\item{alpha}{the elasticnet mixing parameter, see \code{\link[glmnet]{glmnet}}.}

\item{lambda}{optional user supplied lambda sequence, \code{\link[glmnet]{glmnet}}.}

\item{contrasts.arg}{an optional list. See 
the contrasts.arg of
\code{[stats]{model.matrix.default}}.}

\item{xlevs}{a named list of character vectors 
giving the full set of levels to be assumed 
for each factor. See \code{[stats]{model.frame}}.}

\item{na.action}{a function which indicates 
what should happen when the data contain NAs. 
See \code{[stats]{model.frame}}.}

\item{...}{any of the options in \code{\link[glmnet]{glmnet}} or \code{\link[glmnet]{cv.glmnet}}.}
}
\description{
Extends functionality of survival models in \code{\link[glmnet]{glmnet}} to compute survival curves and other calibrations.
}
\details{
This functions offers a user friendly formular-data interface for fitting survival models using \code{\link[glmnet]{glmnet}}. It performs both the cross-validation and model fitting, depending on which option is specified in \code{fittype}. Any additional \code{\link[glmnet]{glmnet}} or \code{\link[glmnet]{cv.glmnet}} arguments can be specified in \code{...}.
}
\examples{

# 
data(veteran, package="survival")
## Fit unpenalized Cox using glmnet
lam <- 0 # Should fit unpenalized Cox model
gfit1 <- glmnetsurv(Surv(time, status) ~ factor(trt) + karno + diagtime + age + prior
	, data = veteran
	, lambda = lam
	, alpha = 1
	, fittype = "fit"
)
print(gfit1)

# Perform cross-validation
gfit2 <- glmnetsurv(Surv(time, status) ~ factor(trt) + karno + diagtime + age + prior
	, data = veteran
	, lambda = NULL
	, alpha = 1
	, fittype = "cv"
)
plot(gfit2)
}
\seealso{
\code{\link[survival]{Surv}}, \code{\link[glmnet]{glmnet}}, \code{\link[glmnet]{cv.glmnet}}
}
