% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnetpostsurv.R
\name{glmnetsurvdata}
\alias{glmnetsurvdata}
\title{Prepare data for glmnet model}
\usage{
glmnetsurvdata(
  formula = formula(data),
  data = sys.parent(),
  contrasts.arg = NULL,
  xlevs = NULL,
  na.action = na.omit
)
}
\arguments{
\item{formula}{Object of class formula describing 
the model. The response and terms are specified 
similar to \code{\link[survival]{Surv}} function.}

\item{data}{optional data frame containing 
variables specified in the formula.}

\item{contrasts.arg}{an optional list. See 
the contrasts.arg of
\code{[stats]{model.matrix.default}}.}

\item{xlevs}{a named list of character vectors 
giving the full set of levels to be assumed 
for each factor. See \code{[stats]{model.frame}}.}

\item{na.action}{a function which indicates 
what should happen when the data contain NAs. 
See \code{[stats]{model.frame}}.}
}
\description{
Prepare data for glmnet model
}
\keyword{internal}
