% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnetsurvfit.R
\name{glmnetsurvfit.glmnet}
\alias{glmnetsurvfit.glmnet}
\alias{glmnetsurvfit}
\alias{glmnetbasehaz.glmnet}
\alias{glmnetbasehaz}
\title{Compute survival curve and cumulative hazard from a glmnet model}
\usage{
\method{glmnetsurvfit}{glmnet}(fit, Y, X, newX, s, wt = NULL, ...)

\method{glmnetbasehaz}{glmnet}(fit, centered = TRUE)
}
\arguments{
\item{fit}{fitted \code{\link[glmnet]{glmnet}} object}

\item{Y}{\code{link[survival]{Surv}} object used in the \code{fit}.}

\item{X}{input matrix used in the \code{fit}.}

\item{newX}{a data frame containing the variables appearing on the right hand side of \code{\link[glmnet]{glmnet}} formula.}

\item{s}{value of \code{lambda} at which predictions are required.}

\item{wt}{option weight applied.}

\item{...}{for future implementations}

\item{centered}{if \code{TRUE} (default), return data from a predicted survival function at the mean values of the predictors, if \code{FALSE} returns prediction for all predictors equal to zero (baseline hazard).}
}
\value{
\code{glmnetsurvfit} and \code{glmnetbasehaz} return S3 objects of class \code{\link[glmnetpostsurv]{glmnetsurvfit.glmnet}} and \code{\link[glmnetpostsurv]{glmnetbasehaz.glmnet}}, respectively:
\item{n}{number of observations used in the fit.}
\item{events}{total number of events of interest in the fit.}
\item{time}{time points defined by the risk set.}
\item{n.risk}{the number of individuals at risk at time \code{t}.}
\item{n.event}{the number of events that occur at time \code{t}.}
\item{n.censor}{the number of subjects who exit the risk set, without an event, at time \code{t}.}
\item{surv}{a vector or a matrix of estimated survival function.}
\item{cumhaz, hazard}{a vector or a matrix of estimated cumulative hazard.}
\item{xmeans}{column means for X.}
\item{s}{lambda value.}
\item{call}{the call that produced the object.}
}
\description{
Compute the predicted survivor and cumulative hazard function for a penalized Cox proportional hazard model.
}
\details{
\code{glmnetsurvfit} and \code{glmnetbasehaz} functions produce survival curves and estimated cumulative hazard, respectively, for the fitted \code{\link[glmnet]{glmnet}} model. They both return the estimated survival probability and the estimated cumulative hazard, which are both Breslow estimate.

The \code{glmnetbasehaz} is an alias for \code{glmnetsurvfit} which simply computed the predicted survival estimates (baseline).

If the \code{newX} argument is missing, the "average" survival or cumulative hazard estimates are produced with the predictor values equal to means of the data set. See \code{\link[survival]{survfit.coxph}} for warning against this. If the \code{newX} is specified, then the returned object will contain a matrix of both survival and cumulative hazard estimates with each column for each row in the \code{newX}.
}
\examples{
data(veteran, package="survival")
lam <- 0
alp <- 1
sobj <- with(veteran, Surv(time, status))
X <- model.matrix(~factor(trt) + karno + diagtime + age + prior, data = veteran)[,-1]
gmodel <- glmnet(X, sobj, 'cox', alpha = alp, lambda = lam)

# Survival estimate
gsurv <- glmnetsurvfit(fit = gmodel, Y = sobj, X = X, s = lam)

# Baseline survival estimate
gbsurv <- glmnetbasehaz(gmodel, centered = FALSE)

}
\seealso{
\code{\link[glmnet]{glmnet}}, \code{\link[glmnetpostsurv]{plot.glmnetsurvfit}}
}
