% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnetsurvfit.R
\name{glmnetsurvfit.glmnetsurv}
\alias{glmnetsurvfit.glmnetsurv}
\alias{glmnetsurvfit}
\alias{glmnetbasehaz.glmnetsurv}
\alias{glmnetbasehaz}
\title{Compute survival curve and cumulative hazard from a glmnet model through glmnetsurv}
\usage{
\method{glmnetsurvfit}{glmnetsurv}(fit, newdata, survfit = FALSE, ...)

\method{glmnetbasehaz}{glmnetsurv}(fit, centered = TRUE, survfit = FALSE)
}
\arguments{
\item{fit}{fitted \code{\link[glmnetsurv]{glmnetsurv}} object}

\item{newdata}{a matrix containing the variables appearing on the right hand side of the model formula of \code{\link[glmnetsurv]{glmnetsurv}} model.}

\item{survfit}{logical value. If \code{TRUE}, survival function estimation is similar to that of \code{\link[survival]{survfit}}, otherwise \code{\link[glmnetsurv]{breslow}} is used.}

\item{...}{for future implementations}

\item{centered}{if \code{TRUE} (default), return data from a predicted survival function at the mean values of the predictors, if \code{FALSE} returns prediction for all predictors equal to zero (baseline hazard).}
}
\value{
\code{glmnetsurvfit} and \code{glmnetbasehaz} return S3 objects of class \code{\link[glmnetsurv]{glmnetsurvfit.glmnetsurv}} and \code{\link[glmnetsurv]{glmnetbasehaz.glmnetsurv}}, respectively:
\item{n}{number of observations used in the fit.}
\item{events}{total number of events of interest in the fit.}
\item{time}{time points defined by the risk set.}
\item{n.risk}{the number of individuals at risk at time \code{t}.}
\item{n.event}{the number of events that occur at time \code{t}.}
\item{n.censor}{the number of subjects who exit the risk set, without an event, at time \code{t}.}
\item{surv}{a vector or a matrix of estimated survival function.}
\item{cumhaz, hazard}{a vector or a matrix of estimated cumulative hazard.}
\item{s}{lambda value.}
\item{call}{the call that produced the object.}
}
\description{
Compute the predicted survivor and cumulative hazard function for a penalized Cox proportional hazard model.
}
\details{
\code{glmnetsurvfit} and \code{glmnetbasehaz} functions produce survival curves and estimated cumulative hazard, respectively, for a \code{\link[glmnet]{glmnet}} model fitted through \code{\link[glmnetsurv]{glmnetsurv}}. They both return the estimated survival probability and the estimated cumulative hazard, which are both Breslow estimate.

The \code{glmnetbasehaz} is an alias for \code{glmnetsurvfit} which simply computed the predicted survival estimates (baseline).

If the \code{newdata} argument is missing, the "average" survival or cumulative hazard estimates are produced with the predictor values equal to means of the data set. See \code{\link[survival]{survfit.coxph}} for warning against this. If the \code{newdata} is specified, then the returned object will contain a matrix of both survival and cumulative hazard estimates with each column for each row in the \code{newdata}.
}
\examples{
data(veteran, package="survival")
lam <- 0
alp <- 1
gmodel <- glmnetsurv(Surv(time, status) ~ factor(trt) + karno + diagtime + age + prior
	, data = veteran
	, alpha = alp
	, lambda = lam
)

# Survival estimate
gsurv <- glmnetsurvfit(fit = gmodel)

# Baseline survival estimate
gbsurv <- glmnetbasehaz(gmodel, centered = FALSE)

}
\seealso{
\code{\link[glmnet]{glmnet}}, \code{\link[glmnetsurv]{plot.glmnetsurvfit}}
}
