% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnetsurvtheme.R
\name{gsurvtheme}
\alias{gsurvtheme}
\title{Set theme for glmnetsurv plots}
\usage{
gsurvtheme()
}
\description{
Sets a theme for glmnetsurv and other ggplot objects
}
\examples{
library(ggplot2)
gsurvtheme()
data(veteran, package="survival")
lam <- 0.02
alp <- 1
gfit <- glmnetsurv(Surv(time, status) ~ factor(trt) + karno + diagtime + age + prior
	, data = veteran
	, lambda = 0
	, alpha = 1
)

# Plot survival curves
gsurv <- glmnetsurvfit(gfit)
plot(gsurv)
}
