% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnetsurvfit.R
\name{newrisk}
\alias{newrisk}
\title{Predict risk for glmnet object}
\usage{
newrisk(model, newdata)
}
\arguments{
\item{newdata}{a matrix containing the variables appearing in model \code{\link[glmnetsurv]{glmnetsurv}} formula.}

\item{object}{fitted \code{\link[glmnetsurv]{glmnetsurv}}.}
}
\description{
Formats newX so that it takes in newdata in the format of the original data. The user's don't have to worry about restructuring the new data.
}
\keyword{internal}
