% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnetsurvfit.R
\name{permuteImp}
\alias{permuteImp}
\title{Permutation variable importance}
\usage{
permuteImp(model, newdata, nrep = 50)
}
\arguments{
\item{model}{fitted \code{\link[glmnetsurv]{glmnetsurv}}, \code{\link[pcoxtime]{pcoxtime}}, \code{\link[survival]{coxph}}, etc.}

\item{newdata}{optional data frame containing the variables appearing on the right hand side of \code{\link[glmnetsurv]{glmnetsurv}} formula.}

\item{nrep}{number of replicates for permulations}
}
\value{
a named vector of variable scores
}
\description{
Computes the relative importance based on random permutation of focal variable for various survival models.
}
\details{
Given predictors \code{x_1, x_2, ..., x_n} used to predict the survival outcome, \code{y}. Suppose, for example, \code{x_1} has low predictive power for the response. Then, if we randomly permute the observed values for \code{x_1}, then the prediction for \code{y} will not change much. Conversely, if any of the predictors highly predicts the response, the permutation of that specific predictor will lead to a considerable change in the predictive measure of the model. In this case, we conclude that this predictor is important. In our implementation, Harrel's concordance index is used to measure the prediction accuracy.
}
\examples{

data(veteran, package="survival")
# Penalized
lam <- 0.1
alp <- 0.5
pfit1 <- glmnetsurv(Surv(time, status) ~ factor(trt) + karno + diagtime + age + prior
	, data = veteran
	, lambda = lam
	, alpha = alp
)
imp <- permuteImp(pfit1, veteran, 10)
imp

}
