% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.glmnetsurv}
\alias{plot.glmnetsurv}
\title{Plot glmnet objects through glmnetsurv}
\usage{
\method{plot}{glmnetsurv}(x, ...)
}
\arguments{
\item{x}{\code{\link[glmnetsurv]{glmnetsurv}} object.}

\item{...}{additional \code{\link[glmnet]{plot.glmnet}} plots methods arguments.}
}
\description{
Plot glmnet objects through glmnetsurv
}
\details{
Internally extract \code{\link[glmnet]{glmnet}} and call plot method
}
