% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postsurvplots.R
\name{plot.glmnetsurvcv}
\alias{plot.glmnetsurvcv}
\title{Plot solution path for glmnetsurvcv}
\usage{
\method{plot}{glmnetsurvcv}(
  x,
  ...,
  type = c("cve", "fit"),
  xvar = c("lambda", "l1"),
  show_nzero = FALSE,
  seed = 1234
)
}
\arguments{
\item{x}{fitted \code{\link[glmnetsurv]{glmnetsurvcv}} object.}

\item{...}{for future implementations}

\item{type}{which plot to return. \code{type = "cve"} (default) return a cross-validation curve and \code{type = "fit"} returns coefficient profiles (solution path). See details.}

\item{xvar}{only if \code{type = "fit"}. Plot coefficients a function of either lambda (\code{xvar = "lambda"}) or l1-norm (\code{xvar = "l1"}).}

\item{show_nzero}{logical. Whether to show number of nonzero coefficients on the plot. Default is \code{show_nzero = FALSE}. Still experimental for \code{type = "cve"}.}

\item{seed}{random number generator. Important if \code{refit = FALSE} in \code{\link[glmnetsurv]{glmnetsurvcv}}.}
}
\value{
a \code{\link[ggplot2]{ggplot}} object.
}
\description{
Plots the cross-validation curve, and upper and lower standard deviation curves, as a function of the optimal lambdas. Also, plots the solution path as a function of optimal lambdas (or randomly picked fold, if \code{refit = FALSE}) or \code{l1}-norm.
}
\details{
To plot solution path corresponding to optimal alpha and lambda, set \code{refit = TRUE} in \code{\link[glmnetsurv]{glmnetsurvcv}}. The plot is a \code{\link[ggplot2]{ggplot}} object, hence can be be customized further.
}
\examples{

library(ggplot2)
data(veteran, package="survival")
# Using a vector of alphas = (0.8, 1)
cv1 <- glmnetsurvcv(Surv(time, status) ~ factor(trt) + karno + diagtime + age + prior
	, data = veteran
	, alpha = c(0.8, 1)
	, refit = TRUE
)
# Plot cross-validation curves
plot(cv1, type = "cve")

# Plot
plot(cv1, type = "fit")

}
