% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postsurvplots.R
\name{plotImp}
\alias{plotImp}
\title{Generic method for plotting variable importance of various models}
\usage{
plotImp(x, drop_zero = FALSE, label_size = 3, ...)
}
\arguments{
\item{x}{a \code{\link[glmnetsurv]{glmnetsurv}} object.}

\item{drop_zero}{if \code{TRUE} only nonzero estimates are shown.}

\item{label_size}{the size of text labels}

\item{...}{for future implementations}
}
\description{
Generic method for plotting variable importance of various models
}
\examples{

data(veteran, package="survival")
# glmnet
gfit1 <- glmnetsurv(Surv(time, status) ~ factor(trt) + karno + diagtime + age + prior
	, data = veteran
	, lambda = 0.02
	, alpha = 0.8
)
imp1 <- varImp(gfit1, type = "param")
plotImp(imp1)
imp2 <- varImp(gfit1, type = "variable", newdata = veteran)
plotImp(imp2)

}
\seealso{
\code{\link[glmnetsurv]{varImp}}
}
