% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnetsurvfit.R
\name{predict.glmnetsurv}
\alias{predict.glmnetsurv}
\title{Prediction for glmnetsurv model}
\usage{
\method{predict}{glmnetsurv}(
  object,
  ...,
  newdata = NULL,
  type = c("lp", "risk", "terms"),
  terms = names(object$assign),
  na.action = na.pass
)
}
\arguments{
\item{object}{fitted \code{\link[glmnetsurv]{glmnetsurv}} object}

\item{...}{for future implementations.}

\item{newdata}{optional data frame containing the variables appearing on the right hand side of \code{\link[glmnetsurv]{glmnetsurv}} formula. If absent, the predictions are for the data frame used in the original fit.}

\item{type}{the type of predicted value. Either linear predictor (\code{"lp"}), the risk score (\code{"risk"} equivalently \code{exp(lp)}) and the terms of linear predictor (\code{"terms"}).}

\item{terms}{if \code{type = "terms"}, this argument can be used to specify which terms to be return. Default is all.}

\item{na.action}{defines the missing value action for the \code{newdata}. If \code{newdata} is absent, then the behavior of missing is dictated by the \code{na.action} option of the original fit.}
}
\value{
a vector of predictions, depending on the \code{type}.
}
\description{
Compute fitted values and model terms for the glmnetsurv model.
}
\details{
The computation of these predictions similar to those in \code{\link[survival]{predict.coxph}}. Our current implementation does not incorporate stratification.
}
\examples{

data(veteran, package="survival")
# Penalized
lam <- 0.02
alp <- 1
gfit1 <- glmnetsurv(Surv(time, status) ~ factor(trt) + karno + diagtime + age + prior
	, data = veteran
	, alpha = alp
	, lambda = lam
)
predict(gfit1)

}
