% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnetsurvfit.R
\name{predictRisk.glmnet}
\alias{predictRisk.glmnet}
\alias{predictRisk}
\title{Extract predictions from glmnet model}
\usage{
predictRisk.glmnet(object, newdata, times, Y, X, s, wt = NULL, ...)
}
\arguments{
\item{object}{fitted \code{\link[glmnet]{glmnet}}.}

\item{newdata}{a matrix containing the variables appearing in X used in fitting \code{\link[glmnet]{glmnet}} model.}

\item{times}{a vector of times in the range of the response, at which to return the survival probabilities.}

\item{Y}{\code{link[survival]{Surv}} object used in the \code{fit}.}

\item{X}{input matrix used in the \code{fit}.}

\item{s}{value of \code{lambda} at which predictions are required.}

\item{wt}{option weight applied.}

\item{...}{for future implementations.}
}
\value{
a matrix of probabilities with as many rows as the rows of the \code{newdata} and as many columns as number of time points (\code{times}).
}
\description{
Extract event probabilities from the fitted model.
}
\details{
For survival outcome, the function predicts the risk, \eqn{1 - S(t|x)}, where \eqn{S(t|x)} is the survival chance of an individual characterized by \eqn{x}.
}
\examples{

data(veteran, package="survival")
# Penalized
lam <- 0.02
alp <- 1
sobj <- with(veteran, Surv(time, status))
X <- model.matrix(~factor(trt) + karno + diagtime + age + prior, data = veteran)[,-1]
gfit1 <- glmnet(X, sobj, 'cox', alpha = alp, lambda = lam)
r1 <- predictRisk.glmnet(gfit1, newdata = X[1:80,]
	, times = 10, sobj, X, lam
)

# Unpenalized model
lam2 <- 0
alp2 <- 1
gfit2 <- glmnet(X, sobj, 'cox', alpha = alp2, lambda = lam2)
r2 <- predictRisk.glmnet(gfit1, newdata = X[1:80,]
	, times = 10, sobj, X, lam2
)
plot(r1, r2, xlim=c(0,1), ylim=c(0,1)
	, xlab = "Penalized predicted survival chance at 10"
	, ylab="Unpenalized predicted survival chance at 10"
)

}
