% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnetsurvfit.R
\name{predictRisk.glmnetsurv}
\alias{predictRisk.glmnetsurv}
\alias{predictRisk}
\title{Extract predictions from glmnet model}
\usage{
\method{predictRisk}{glmnetsurv}(object, newdata, times, ...)
}
\arguments{
\item{object}{fitted \code{\link[glmnetsurv]{glmnetsurv}}.}

\item{newdata}{a matrix containing the variables appearing in model \code{\link[glmnetsurv]{glmnetsurv}} formula.}

\item{times}{a vector of times in the range of the response, at which to return the survival probabilities.}

\item{...}{for future implementations.}
}
\value{
a matrix of probabilities with as many rows as the rows of the \code{newdata} and as many columns as number of time points (\code{times}).
}
\description{
Extract event probabilities from the fitted model.
}
\details{
For survival outcome, the function predicts the risk, \eqn{1 - S(t|x)}, where \eqn{S(t|x)} is the survival chance of an individual characterized by \eqn{x}.
}
\examples{

data(veteran, package="survival")
# Penalized
lam <- 0.02
alp <- 1
gfit1 <- glmnetsurv(Surv(time, status) ~ factor(trt) + karno + diagtime + age + prior
	, data = veteran
	, alpha = alp
	, lambda = lam
)
r1 <- predictRisk(gfit1, newdata = veteran[1:80,], times = 10)

# Unpenalized model
lam2 <- 0
alp2 <- 1
gfit2 <- glmnetsurv(Surv(time, status) ~ factor(trt) + karno + diagtime + age + prior
	, data = veteran
	, alpha = alp2
	, lambda = lam2
)
r2 <- predictRisk(gfit2, newdata = veteran[1:80,], times = 10)
plot(r1, r2, xlim=c(0,1), ylim=c(0,1)
	, xlab = "Penalized predicted survival chance at 10"
	, ylab="Unpenalized predicted survival chance at 10"
)

}
