% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.glmnetsurv}
\alias{print.glmnetsurv}
\title{Print method fot  glmnetsurv object}
\usage{
\method{print}{glmnetsurv}(x, ...)
}
\arguments{
\item{x}{fitted \code{\link[glmnetsurv]{glmnetsurv}} model object}

\item{...}{for future implementations}
}
\value{
see \code{\link[glmnet]{glmnet}}.
}
\description{
This function prints a summary of the glmnet object from  glmnetsurv.
}
\details{
The call that produced \code{\link[glmnetsurv]{glmnetsurv}} is printed, followed by \code{\link[glmnet]{glmnet}} objects.
}
