% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.glmnetsurvcv}
\alias{print.glmnetsurvcv}
\title{Print cross-validated glmnetsurvcv object}
\usage{
\method{print}{glmnetsurvcv}(x, ...)
}
\arguments{
\item{x}{\code{\link[glmnetsurv]{glmnetsurvcv}} object}

\item{...}{for future implementations}
}
\description{
Print the summary of the result of cross-validation for a glmnetsurv object.
}
\details{
A summary of optimal lambda and alpha for training glmnetsurv model.
}
