% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.glmnetsurvfit}
\alias{print.glmnetsurvfit}
\title{Print a short summary of survival function}
\usage{
\method{print}{glmnetsurvfit}(x, ...)
}
\arguments{
\item{x}{the result of a call to the \code{\link[glmnetsurv]{glmnetsurvfit.glmnetsurv}} function.}

\item{...}{for future implementations}
}
\value{
The call to the \code{\link[glmnetsurv]{glmnetsurvfit.glmnetsurv}} and the summary of the survival function.
}
\description{
Print the number of observations and number of events.
}
\details{
Provide a summary of \code{\link[glmnetsurv]{glmnetsurvfit.glmnetsurv}} object.
}
