% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnetHazard.R
\name{robustHazard}
\alias{robustHazard}
\title{Breslow estimator for the individuals at risk}
\usage{
robustHazard(y, lp)
}
\arguments{
\item{y}{\code{\link[survival]{Surv}} object.}

\item{lp}{a vector of linear predictor. Computed from the \code{X} matrix (predictors) and the estimated coefficients (\code{beta}).}
}
\value{
A list of S3 objects. 
\item{n}{number of observations used in the fit.}
\item{time}{time points defined by the risk set.}
\item{n.event}{the number of events that occur at time \code{t}.}
\item{n.risk}{the number of individuals at risk at time \code{t}.}
\item{n.censor}{the number of subjects who exit the risk set, without an event, at time \code{t}.}
}
\description{
Computes the number of indiduals ar risk given the linear predictor
}
\keyword{internal}
