% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnetpostsurv.R
\name{varImp.glmnetsurv}
\alias{varImp.glmnetsurv}
\alias{varImp}
\title{Compute variable importance of glmnetsurv object}
\usage{
\method{varImp}{glmnetsurv}(object, show_sign = FALSE, scale = FALSE, ...)
}
\arguments{
\item{object}{fitted \code{\link[glmnetsurv]{glmnetsurv}} object.}

\item{show_sign}{if \code{FALSE}, the absolute value of coefficients are used. If \code{TRUE} estimated model estimated coefficients are used.}

\item{scale}{if \code{TRUE} the scores are divided by the absolute sum of the coefficients.}

\item{...}{not implemented.}
}
\description{
Compute variable importance of glmnetsurv object
}
\details{
If \code{show_sign = FALSE} the absolute value of the coefficients corresponding the tuned model are used otherwise, the estimated coefficients are used.
}
\examples{

data(veteran, package="survival")
# glmnet
gfit1 <- glmnetsurv(Surv(time, status) ~ factor(trt) + karno + diagtime + age + prior
	, data = veteran
	, lambda = 0.02
	, alpha = 0.8
)
imp1 <- varImp(gfit1)
print(imp1)
imp2 <- varImp(gfit1, show_sign = TRUE, scale = TRUE)
print(imp2)

}
\seealso{
\code{\link[glmnetsurv]{plot.varImp}}
}
