% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnetpostsurv.R
\name{varImp}
\alias{varImp}
\title{Compute variable importance of various survival models object}
\usage{
varImp(
  object,
  type = c("param", "variable"),
  scale = TRUE,
  newdata,
  nrep = 20,
  ...
)
}
\arguments{
\item{object}{fitted \code{\link[glmnetsurv]{glmnetsurv}}, \code{\link[pcoxtime]{pcoxtime}}, \code{\link[survival]{coxph}}, etc, object.}

\item{type}{if \code{type = "param"} absolute value of estimated coefficients are used. If \code{type = "variable"} variable level importance is computed using permutation.}

\item{scale}{if \code{TRUE} the scores are divided by the absolute sum of the coefficients.}

\item{newdata}{optional data frame containing the variables appearing on the right hand side of \code{\link[glmnetsurv]{glmnetsurv}} formula. Required if \code{type = "variable"}}

\item{nrep}{number of replicates for permulations. Only if \code{type = "variable"}.}

\item{...}{not implemented.}
}
\description{
Compute variable importance of various survival models object
}
\details{
Absolute value of the coefficients (parameters) corresponding the tuned model are used \code{type = param}. Otherwise, variable level importance is computed using permutation. See  \code{\link[glmnetsurv]{permuteImp}}.
}
\examples{

data(veteran, package="survival")
# glmnet
gfit1 <- glmnetsurv(Surv(time, status) ~ factor(trt) + karno + diagtime + age + prior
	, data = veteran
	, lambda = 0.02
	, alpha = 0.8
)
imp1 <- varImp(gfit1, type = "param")
print(imp1)
imp2 <- varImp(gfit1, type = "variable", newdata = veteran)
print(imp2)

}
\seealso{
\code{\link[glmnetsurv]{plotImp}}
}
