% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postsurvplots.R
\name{plot.glmnetsurvfit}
\alias{plot.glmnetsurvfit}
\title{Plot survival and cumulative hazard curves}
\usage{
\method{plot}{glmnetsurvfit}(x, ..., type = c("surv", "cumhaz"), lsize = 0.3, compare = FALSE)
}
\arguments{
\item{x}{a \code{\link[glmnetsurv]{glmnetsurvfit.glmnetsurv}} or \code{\link[glmnetsurv]{glmnetbasehaz.glmnetsurv}} object.}

\item{...}{for future implementations}

\item{type}{type of curve to generate. Either \code{type = "surv"} for survival curves or \code{type = "cumhaz"} for cumulative hazard curve.}

\item{lsize}{line size for the curves. Default is \code{0.3}.}

\item{compare}{logical. Whether to return plot with labels to add additional \code{geom} object for comparison. Default is \code{FALSE}.}
}
\value{
a \code{\link[ggplot2]{ggplot}} object.
}
\description{
Plot estimated survival and cumulative  hazard curves for \code{glmnet} model.
\code{\link[glmnetsurv]{glmnetsurvfit.glmnetsurv}}
}
\details{
Depending on the specification in \code{\link[glmnetsurv]{glmnetsurvfit.glmnetsurv}}, this function plots either average or individual survival or cumulative hazard curves. The plot is a \code{\link[ggplot2]{ggplot}} object, hence can be be customized further, see example below.
}
\examples{

library(ggplot2)
data(veteran, package="survival")

gmodel <- glmnetsurv(Surv(time, status) ~ factor(trt) + karno + diagtime + age + prior
	, data = veteran
	, lambda = 0
	, alpha = 1
)

# Survival estimate
gsurv <- glmnetsurvfit(fit = gmodel)

# Plot survival curves
plot(gsurv)

# Baseline survival estimate
gbsurv <- glmnetbasehaz(gmodel, centered = FALSE)
plot(gbsurv)

# Compare overall and baseline cumulative hazard
p1 <- plot(gsurv, type = "cumhaz", compare = TRUE)
df2 <- data.frame(time = gbsurv$time, cumhaz = gbsurv$hazard)
p2 <- (p1
	+ geom_line(data = df2, aes(x = time, y = cumhaz, group = 1, col = "baseline"))
	+ scale_colour_manual(name = "C. hazard"
		, values = c("#E41A1C", "#000000")
		, labels = c("baseline", "overall")
	)
)
print(p2)

}
